+++ modules/users/mappers/UserMapper.ts
@@ -61,6 +61,7 @@ type PrismaUserRelations = Pick<
   lastVersionCode?: number | null;
   lastVersionName?: string | null;
   lastVersionUpdate?: Date | null;
+  lastOtaUpdateId?: string | null;
   lastLoginAt?: Date | null;
   bankName?: string | null;
   bankAccountNo?: string | null;